update to vs17, everything is broken

+++ src/platform_win32.cpp
@@ -470,7 +470,7 @@ void* platform_heap_alloc(u64 bytes) {
 	// cout << "alloc" << endl;
 
 	HANDLE heap = GetProcessHeap();
-	return HeapAlloc(heap, HEAP_ZERO_MEMORY | HEAP_GENERATE_EXCEPTIONS, bytes);
+	return HeapAlloc(heap, HEAP_ZERO_MEMORY | HEAP_GENERATE_EXCEPTIONS, (SIZE_T)bytes);
 }
 
 void platform_heap_free(void* mem) {
